+++ example/lib/main.dart
@@ -203,7 +203,7 @@ class _MyAppState extends State<MyApp> {
           this.loading = false;
           this.albums = albums;
           if (this.albums != null && this.albums!.isNotEmpty)
-            this.currentAlbum = albums.first;
+            this.currentAlbum = albums!.first;
           if (controller != null && currentAlbum != null)
             this.controller!.reloadAlbum(currentAlbum!.identifier);
         });