+++ lib/plugin.dart
@@ -6,7 +6,7 @@ import 'package:image_list/data/media.dart';
 class ImageListPlugin {
   static const MethodChannel _channel = const MethodChannel('image_list');
 
-  static Future<dynamic> getAlbums({
+  static Future<List<Album>?> getAlbums({
     List<MediaType> types = const <MediaType>[MediaType.image, MediaType.video],
   }) async {
     bool hasPermission = await checkPermission();
@@ -53,6 +53,27 @@ class ImageListPlugin {
 
     return Uint8List.fromList(raw ?? <int>[]);
   }
+
+  static Future<Uint8List> getAlbumThumbnail({
+    required String albumUri,
+    int? width,
+    int? height,
+    int? size,
+    int quality = 100,
+  }) async {
+    final raw = await _channel.invokeMethod(
+      'getAlbumThumbnail',
+      {
+        'albumUri': albumUri,
+        'width': width,
+        'height': height,
+        'size': size,
+        'quality': quality,
+      },
+    );
+
+    return Uint8List.fromList(raw ?? <int>[]);
+  }
 }
 
 class Album {
@@ -85,5 +106,26 @@ class Album {
   @override
   String toString() {
     return "Album(name: $name, identifier: $identifier, count: $count)";
+  }
+}
+
+class AlbumWithThumbnail extends Album {
+  final Uint8List thumbnail;
+
+  AlbumWithThumbnail(
+    String name,
+    String identifier,
+    int count, {
+    Uint8List? thumbnail,
+  })  : this.thumbnail = thumbnail ?? Uint8List.fromList([]),
+        super(name, identifier, count);
+
+  static AlbumWithThumbnail fromAlbum(Album album, {Uint8List? thumbnail}) {
+    return AlbumWithThumbnail(
+      album.name,
+      album.identifier,
+      album.count,
+      thumbnail: thumbnail,
+    );
   }
 }